requirements.txt
streamlit
plyer
python-dateutil
pytz
ics
underthesea